+++ finagle-base-http/src/test/scala/com/twitter/finagle/http/DefaultHeaderMapTest.scala
@@ -69,4 +69,28 @@ class DefaultHeaderMapTest extends AbstractHeaderMapTest with GeneratorDrivenPro
       intercept[IllegalArgumentException](DefaultHeaderMap(h))
     }
   }
+
+  test("does not validate header names or values with addUnsafe") {
+    val headerMap = newHeaderMap()
+
+    forAll(genInvalidHeaderName) { h =>
+      headerMap.addUnsafe(h._1, h._2)
+    }
+
+    forAll(genInvalidHeaderValue) { h =>
+      headerMap.addUnsafe(h._1, h._2)
+    }
+  }
+
+  test("does not validate header names or values with setUnsafe") {
+    val headerMap = newHeaderMap()
+
+    forAll(genInvalidHeaderName) { h =>
+      headerMap.setUnsafe(h._1, h._2)
+    }
+
+    forAll(genInvalidHeaderValue) { h =>
+      headerMap.setUnsafe(h._1, h._2)
+    }
+  }
 }